requirements.txt
praw
openai
tqdm
python-dotenv
